[gl] Moved GLObjectStore to gl namespace
