+++ src/mbgl/geometry/buffer.hpp
@@ -3,7 +3,7 @@
 
 #include <mbgl/gl/gl.hpp>
 #include <mbgl/platform/log.hpp>
-#include <mbgl/util/gl_object_store.hpp>
+#include <mbgl/gl/gl_object_store.hpp>
 #include <mbgl/util/noncopyable.hpp>
 #include <mbgl/util/thread_context.hpp>
