+++ src/mbgl/gl/gl_object_store.hpp
@@ -4,11 +4,10 @@
 #include <mbgl/gl/gl.hpp>
 #include <mbgl/util/noncopyable.hpp>
 
-#include <cstdint>
 #include <vector>
 
 namespace mbgl {
-namespace util {
+namespace gl {
 
 class GLObjectStore : private util::noncopyable {
 public:
@@ -29,7 +28,7 @@ private:
     std::vector<GLuint> abandonedTextures;
 };
 
-} // namespace util
+} // namespace gl
 } // namespace mbgl
 
 #endif
