+++ src/mbgl/util/texture_pool.cpp
@@ -1,6 +1,6 @@
 #include <mbgl/util/texture_pool.hpp>
 
-#include <mbgl/util/gl_object_store.hpp>
+#include <mbgl/gl/gl_object_store.hpp>
 #include <mbgl/util/thread_context.hpp>
 
 #include <vector>
